+++ src/mock-builders/api/index.js
@@ -32,5 +32,6 @@ export const useMockedApis = (axios, apiResponses) => {
 export * from './queryChannels';
 export * from './queryMembers';
 export * from './getOrCreateChannel';
+export * from './threadReplies';
 export * from './sendMessage';
 export * from './error';