+++ types/index.d.ts
@@ -150,7 +150,7 @@ export interface ChannelListProps extends ChatContextValue {
   LoadingErrorIndicator?: React.ElementType<LoadingErrorIndicatorProps>;
   List?: React.ElementType<ChannelListUIComponentProps>;
   Paginator?: React.ElementType<PaginatorProps>;
-
+  lockChannelOrder?: boolean;
   onMessageNew?(
     thisArg: React.Component<ChannelListProps>,
     e: Client.Event<Client.MessageNewEvent>,
@@ -293,7 +293,7 @@ export interface LoadingErrorIndicatorProps extends TranslationContextValue {
 
 export interface AvatarProps {
   /** image url */
-  image?: string;
+  image?: string | null;
   /** name of the picture, used for title tag fallback */
   name?: string;
   /** shape of the avatar, circle, rounded or square */
@@ -334,6 +334,7 @@ export interface MessageListProps
   HeaderComponent?: React.ElementType;
   /** Component to render at the top of the MessageList */
   EmptyStateIndicator?: React.ElementType<EmptyStateIndicatorProps>;
+  LoadingIndicator?: React.ElementType<LoadingIndicatorProps>;
   /** Date separator component to render  */
   dateSeparator?: React.ElementType<DateSeparatorProps>;
   /** Turn off grouping of messages by user */
@@ -352,10 +353,7 @@ export interface MessageListProps
   additionalMessageInputProps?: object;
 }
 
-export interface ChannelHeaderProps
-  extends ChannelContextValue,
-    TranslationContextValue,
-    ChatContextValue {
+export interface ChannelHeaderProps {
   /** Set title manually */
   title?: string;
   /** Show a little indicator that the channel is live right now */
@@ -478,6 +476,7 @@ export interface AttachmentUIComponentProps {
 
 // MessageProps are all props shared between the Message component and the Message UI components (e.g. MessageSimple)
 export interface MessageProps extends TranslationContextValue {
+  addNotification?(notificationText: string, type: string): any;
   /** The message object */
   message?: Client.MessageResponse;
   /** The client connection object for connecting to Stream */
@@ -536,7 +535,6 @@ export interface MessageComponentProps
   onUserHover?(e: React.MouseEvent, user: Client.User): void;
   messageActions?: Array<string>;
   members?: SeamlessImmutable.Immutable<{ [user_id: string]: Client.Member }>;
-  addNotification?(notificationText: string, type: string): any;
   retrySendMessage?(message: Client.Message): void;
   removeMessage?(updatedMessage: Client.MessageResponse): void;
   mutes?: Client.Mute[];
@@ -736,9 +734,9 @@ export interface ChatDownProps extends TranslationContextValue {
 
 export interface CommandItemProps {
   entity: {
-    name: string;
-    args: string;
-    description: string;
+    name?: string | null;
+    args?: string | null;
+    description?: string | null;
   };
 }
 
@@ -754,9 +752,9 @@ export interface EmoticonItemProps {
 
 export interface UserItemProps {
   entity: {
-    name: string;
-    id: string;
-    image: string;
+    name?: string | null;
+    id?: string | null;
+    image?: string | null;
   };
 }
 
@@ -769,9 +767,16 @@ export interface GalleryProps {
 }
 
 export interface ImageProps {
-  image_url: string;
-  thumb_url: string;
-  fallback: string;
+  image_url?: string;
+  thumb_url?: string;
+  fallback?: string;
+}
+
+export interface ModalWrapperProps {
+  images: { src: string; source: string }[];
+  toggleModal: (selectedIndex?: number) => void;
+  index?: number;
+  modalIsOpen: boolean;
 }
 
 export interface InfiniteScrollProps {
@@ -786,6 +791,32 @@ export interface InfiniteScrollProps {
   element?: React.ElementType;
   loader?: React.ReactNode;
   threshold?: number;
+  children?: any;
+  listenToScroll?: (offset: number, reverseOffset: number) => void;
+}
+
+export interface ModalImageProps {
+  data: { src: string };
+}
+
+export interface ReverseInfiniteScrollProps {
+  loadMore(): any;
+  hasMore?: boolean;
+  initialLoad?: boolean;
+  isReverse?: boolean;
+  pageStart?: number;
+  isLoading?: boolean;
+  useCapture?: boolean;
+  useWindow?: boolean;
+  element?: React.ElementType;
+  loader?: React.ReactNode;
+  threshold?: number;
+  className?: string;
+  /** The function is called when the list scrolls */
+  listenToScroll?(
+    standardOffset: string | number,
+    reverseOffset: string | number,
+  ): any;
   listenToScroll?(standardOffset: number, reverseOffset: number): void;
   [elementAttribute: string]: any; // any other prop is applied as attribute to element
 }
@@ -821,7 +852,7 @@ export interface MessageRepliesCountButtonProps
   onClick?: React.MouseEventHandler;
 }
 export interface ModalProps {
-  onClose(): void;
+  onClose?(): void;
   open: boolean;
 }
 export interface SafeAnchorProps {}
@@ -860,9 +891,10 @@ export class EditMessageForm extends React.PureComponent<
 > {}
 export const EmoticonItem: React.FC<EmoticonItemProps>;
 export const EmptyStateIndicator: React.FC<EmptyStateIndicatorProps>;
+export const Gallery: React.FC<GalleryProps>;
+export const Image: React.FC<ImageProps>;
+export const ImageModal: React.FC<ModalWrapperProps>;
 export const EventComponent: React.FC<EventComponentProps>;
-export class Gallery extends React.PureComponent<GalleryProps, any> {}
-export class Image extends React.PureComponent<ImageProps, any> {}
 export class InfiniteScroll extends React.PureComponent<
   InfiniteScrollProps,
   any
@@ -879,6 +911,8 @@ export class MessageActionsBox extends React.PureComponent<
 export const MessageNotification: React.FC<MessageNotificationProps>;
 export const MessageRepliesCountButton: React.FC<MessageRepliesCountButtonProps>;
 export class Modal extends React.PureComponent<ModalProps, any> {}
+export const ModalImage: React.FC<ModalImageProps>;
+
 export class ReverseInfiniteScroll extends React.PureComponent<
   InfiniteScrollProps,
   any
@@ -946,44 +980,93 @@ export class MessageCommerce extends React.PureComponent<
 > {}
 
 export interface MessageLivestreamProps extends MessageUIComponentProps {}
-export type MessageLivestreamState = {
-  actionsBoxOpen: boolean;
-  reactionSelectorOpen: boolean;
-};
-export class MessageLivestream extends React.PureComponent<
-  MessageLivestreamProps,
-  MessageLivestreamState
-> {}
+export interface MessageLivestreamActionProps {
+  initialMessage?: boolean;
+  message?: Client.MessageResponse;
+  tDateTimeParser?(datetime: string | number): Dayjs.Dayjs;
+  channelConfig?: Client.ChannelConfig;
+  threadList?: boolean;
+  handleOpenThread?(event: React.BaseSyntheticEvent): void;
+  onReactionListClick?: () => void;
+  getMessageActions(): Array<string>;
+  messageWrapperRef?: React.RefObject<HTMLElement>;
+  setEditingState?(message: Client.MessageResponse): any;
+}
+export const MessageLivestream: React.FC<MessageLivestreamProps>;
 export type MessageTeamState = {
   actionsBoxOpen: boolean;
   reactionSelectorOpen: boolean;
 };
 export interface MessageTeamProps extends MessageUIComponentProps {}
+export interface MessageTeamAttachmentsProps {
+  Attachment?: React.ElementType<AttachmentUIComponentProps>;
+  message?: Client.MessageResponse;
+  handleAction?(
+    name: string,
+    value: string,
+    event: React.BaseSyntheticEvent,
+  ): void;
+}
+export interface MessageTeamStatusProps {
+  t?: i18next.TFunction;
+  threadList?: boolean;
+  lastReceivedId?: string | null;
+  message?: Client.MessageResponse;
+  readBy?: Array<Client.UserResponse>;
+}
 export class MessageTeam extends React.PureComponent<
   MessageUIComponentProps,
   MessageTeamState
 > {}
 
 export interface MessageSimpleProps extends MessageUIComponentProps {}
-export interface MessageSimpleTextProps extends MessageSimpleProps {
-  actionsBoxOpen: boolean;
-  onReactionListClick: () => void;
-  showDetailedReactions: boolean;
-  reactionSelectorRef: React.RefObject<ReactionSelector>;
-  setActionsBoxOpen: (state: boolean) => void;
-  hideOptions: () => void;
+export interface MessageTextProps extends MessageSimpleProps {
+  customOptionProps?: Partial<MessageOptionsProps>;
+  customInnerClass?: string;
+  customWrapperClass?: string;
+  onReactionListClick?: () => void;
+  theme?: string;
+  showDetailedReactions?: boolean;
+  messageWrapperRef?: React.RefObject<HTMLElement>;
 }
-export interface MessageSimpleOptionsProps extends MessageSimpleProps {
-  actionsBoxOpen: boolean;
-  hideOptions: () => void;
-  setActionsBoxOpen: (state: boolean) => void;
-  onReactionListClick: () => void;
-}
-export interface MessageSimpleActionsProps extends MessageSimpleProps {
+
+export interface MessageActionsProps {
   addNotification?(notificationText: string, type: string): any;
-  actionsBoxOpen: boolean;
-  hideOptions: () => void;
-  setActionsBoxOpen: (state: boolean) => void;
+  handleEdit?(event?: React.BaseSyntheticEvent): void;
+  handleDelete?(event?: React.BaseSyntheticEvent): void;
+  handleFlag?(event?: React.BaseSyntheticEvent): void;
+  handleMute?(event?: React.BaseSyntheticEvent): void;
+  mutes?: Client.Mute[];
+  getMessageActions(): Array<string>;
+  getFlagMessageSuccessNotification?(message: MessageResponse): string;
+  getFlagMessageErrorNotification?(message: MessageResponse): string;
+  getMuteUserSuccessNotification?(message: MessageResponse): string;
+  getMuteUserErrorNotification?(message: MessageResponse): string;
+  setEditingState?(message: Client.MessageResponse): any;
+  messageListRect?: DOMRect;
+  message?: Client.MessageResponse;
+  messageWrapperRef?: React.RefObject<HTMLElement>;
+  inline?: boolean;
+  customWrapperClass?: string;
+}
+export interface MessageActionsWrapperProps {
+  customWrapperClass?: string;
+  inline?: boolean;
+  setActionsBoxOpen: (actionsBoxOpen: boolean) => void;
+}
+
+export interface MessageOptionsProps {
+  getMessageActions(): Array<string>;
+  handleOpenThread?(event: React.BaseSyntheticEvent): void;
+  initialMessage?: boolean;
+  message?: Client.MessageResponse;
+  messageWrapperRef?: React.RefObject<HTMLElement>;
+  onReactionListClick?: () => void;
+  threadList?: boolean;
+  displayLeft?: boolean;
+  displayReplies?: boolean;
+  displayActions?: boolean;
+  theme?: string;
 }
 
 export const MessageSimple: React.FC<MessageSimpleProps>;